requirements.txt
django
djangorestframework
django-simple-history
django-extensions
djangorestframework-simplejwt
django-debug-toolbar
django-storages
boto3